feat(www): update app-sign-in to check for pending app sign-in on native-app message
